# benchmark: return time series with missing periods filled in