Previous

```
unix_ts,avg_latency,throughput
1460436229,1s,1
1460436234,1s,1
```

After

```
unix_ts,avg_latency,throughput
1460436229,1s,1
1460436230,0,0
1460436231,0,0
1460436232,0,0
1460436233,0,0
1460436234,1s,1
```